Fix architecture in audit rules
Fix grub symlinks

Change exit code to 1 on error

Use FOLDER variable

Add service stop/start when chkconfiging

fix find to ignore /proc

Fix sed for nullok in system-auth

+++ packages/aqueduct/aqueduct/compliance/Bash/SSG/dev/audit_file_deletions.sh
@@ -20,5 +20,7 @@ set -e
 
 . $(dirname $0)/audit_rules_common
 
-add_rule '-a always,exit -F arch=b64 -S unlink -S unlinkat -S rename -S renameat -F auid>=500 -F auid!=4294967295 -k delete'
-
+add_rule '-a always,exit -F arch=b32 -S unlink -S unlinkat -S rename -S renameat -F auid>=500 -F auid!=4294967295 -k delete'
+if /bin/uname -m|/bin/grep -q 64; then
+    add_rule '-a always,exit -F arch=b64 -S unlink -S unlinkat -S rename -S renameat -F auid>=500 -F auid!=4294967295 -k delete'
+fi
